refactor(bytecode): rename vihaco_isa module to isa (#769)
