+++ crates/bloqade-lanes-bytecode-core/src/isa/validate.rs
